Move Stageable in lib.Misc
