+++ lib/src/main/scala/spinal/lib/Misc.scala
@@ -131,3 +131,22 @@ class FlowCmdRsp[T <: Data, T2 <: Data](cmdType : HardType[T], rspType : HardTyp
     case 1 => RegInit(False) setWhen(cmd.valid) clearWhen(rsp.valid)
   }
 }
+
+
+object Stageable{
+  def apply[T <: Data](gen : => T) = new Stageable(gen)
+  def apply[T <: Data](gen : HardType[T]) = new Stageable(gen.craft())
+}
+
+class Stageable[T <: Data](gen : => T) extends HardType(gen) with Nameable
+
+case class StageableOffset(val value : Any)
+object StageableOffsetNone extends StageableOffset(null)
+case class StageableKey(stageable: Stageable[Data], key : Any){
+  override def toString = {
+    var name = stageable.getName()
+    if(key != null) name = name + "_" + key
+    name
+  }
+}
+
